requirements.txt
asgiref==3.8.1
Django==5.1.4
django-cors-headers==4.6.0
django-filter==24.3
django-jazzmin==3.0.1
djangorestframework==3.15.2
djangorestframework_simplejwt==5.4.0
drf-yasg==1.21.8
inflection==0.5.1
Markdown==3.7
packaging==24.2
pillow==11.1.0
PyJWT==2.10.1
pytz==2024.2
PyYAML==6.0.2
shortuuid==1.0.13
sqlparse==0.5.3
uritemplate==4.1.1
